refactor: 💡 move ValidResult type to types
